Implementando o controle de exceções no resource Address

+++ src/main/java/br/com/postech30/challenge/service/impl/AddressServiceImpl.java
@@ -2,10 +2,12 @@ package br.com.postech30.challenge.service.impl;
 
 import br.com.postech30.challenge.dto.AddressDTO;
 import br.com.postech30.challenge.entity.Address;
+import br.com.postech30.challenge.exceptions.ResourceNotFoundException;
 import br.com.postech30.challenge.repository.AddressRepository;
 import br.com.postech30.challenge.service.AddressService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -35,20 +37,29 @@ public class AddressServiceImpl implements AddressService {
     @Override
     @Transactional(readOnly = true)
     public AddressDTO findById(Long id) {
-        Optional<Address> address = repository.findById(id);
+        Address address = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Endereço não encontrado"));
 
-        return new AddressDTO(address.get());
+        return new AddressDTO(address);
     }
 
     @Override
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Endereço não encontrado");
+        }
+
         repository.deleteById(id);
     }
 
     @Override
     @Transactional
     public void update(Long id, AddressDTO addressDTO) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Endereço não encontrado");
+        }
+
         Address address = repository.getReferenceById(id);
         address = mapTo(addressDTO, address);
         repository.save(address);
